requirements.txt
numpy>=1.23.5
scipy>=1.9.3
pandas>=1.5.2
scikit_learn>=1.2.0
pyyaml>=6.0
colorlog>=4.7.2
colorama>=0.4.6
pymoo>=0.6.0
cvxpy>=1.2.3
numba>=0.56.4